Update ft_strncmp.c
casting num to a int

+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: gpacheco <gpacheco@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/26 14:16:57 by gpacheco          #+#    #+#             */
-/*   Updated: 2021/08/26 14:55:53 by gpacheco         ###   ########.fr       */
+/*   Updated: 2021/08/26 16:43:14 by gpacheco         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int ft_strncmp ( const char * str1, const char * str2, size_t num )
 	{
 		if(str1[i] == str2[j])
 		{
-			if(j < num-1)
+			if(j < (int)num-1)
 				j++;
 			else
 				return (0);
